add comments + error on illegal access(create/delete whiteboards)

+++ OnlineCollaborationPlatform/protected/controllers/WhiteboardController.php
@@ -2,41 +2,67 @@
 
 class WhiteboardController extends Controller
 {
-	public function actionIndex()
+	public function actionList()
 	{
-		$whiteboards = Whiteboard::model()->findAllByAttributes(array('owner'=>Yii::app()->user->id));
-		$this->render('//whiteboard/index',array('whiteboards'=>$whiteboards));
-
+		if (Yii::app() -> user -> isGuest){
+			throw new CHttpException(403,'You are not allowed to list whiteboards. Please login to use this feature.');
+		}else{
+			// Alle Whiteboards, die dem eingeloggten User gehören aus der Datenbank holen.
+			$whiteboards = Whiteboard::model()->findAllByAttributes(array('owner'=>Yii::app()->user->id));
+			$this->render('//whiteboard/list',array('whiteboards'=>$whiteboards));
+		}
 	}
 	
 	public function actionCreate()
 	{
 		if (Yii::app() -> user -> isGuest){
-			
+			throw new CHttpException(403,'You are not allowed to create a whiteboard. Please login to use this feature.');
 		}else{
+			// Whiteboard wird erzeugt und mit Werten befüllt.
 			$whiteboard = new Whiteboard;
 			$whiteboard->name='Whiteboard01'; 
 			$whiteboard->date=date(DATE_RFC822); 
 			$whiteboard->owner=Yii::app()->user->id;
+			
+			// Erzeugtes Whiteboard in der Datenbank ablegen.
 			$whiteboard->save(); 
-			$this->render('//whiteboard/view',array("model"=>$whiteboard));
+			
+			// Weiterleiten um das Whiteboard nach dem Erstellen anzuzeigen.
+			$this->redirect(array('whiteboard/view','id'=>$whiteboard->id));
+			// $this->render('//whiteboard/view',array("model"=>$whiteboard));
 		}
 	}
 	public function actionView($id){
+		// Whiteboard mit 'Primary Key' aus der Datenbank lesen
 		$whiteboard = Whiteboard::model()->findByPK($id);
 		if(!$whiteboard){
 			throw new CHttpException(404,'The specified whiteboard cannot be found.');
+		}else{
+			$this->render('//whiteboard/view',array('whiteboard'=>$whiteboard));
 		}
-		$this->render('//whiteboard/view',array("model" =>$whiteboard));
 	}
 
 	public function actionDelete($id){
-		$whiteboard = Whiteboard::model()->findByPK($id);	
-		if($whiteboard){
-			$whiteboard->delete();
-			$this->redirect(array('whiteboard/index'));
-		}else{
-			throw new CHttpException(404,'The specified whiteboard cannot be found.');
+		// Nur User dürfen Whiteboards löschen
+		if (Yii::app() -> user -> isGuest){
+			throw new CHttpException(403,'You are not allowed to delete a whiteboard. Please login to use this feature.');
+		}
+		else{
+			// Whiteboard aus der Datenbank lesen
+			$whiteboard = Whiteboard::model()->findByPK($id);
+			
+			if($whiteboard){
+				if($whiteboard->owner == Yii::app()->user->id){
+					$whiteboard->delete();
+					$this->redirect(array('whiteboard/list'));
+				}
+				else{
+					throw new CHttpException(403,'You are not the owner. You are not allowed to delete this whiteboard.');
+				}
+			}
+			else{
+				throw new CHttpException(404,'The specified whiteboard cannot be found.');
+			}
 		}
 	}
 	// Uncomment the following methods and override them if needed
